+++ ext/rbkit_event_packer.h
@@ -1,8 +1,14 @@
 #ifndef RBKIT_MESSAGE_PACKER
 #define RBKIT_MESSAGE_PACKER
-#define RBKIT_PROTOCOL_VERSION "1.0"
+#define RBKIT_PROTOCOL_VERSION "2.0"
 #include "msgpack.h"
 #include "rbkit_event.h"
+
+// Object dump will be split into multiple
+// messages. This macro defines the number
+// of object data that should be packed
+// as the payload of one message.
+#define MAX_OBJECT_DUMPS_IN_MESSAGE 1000
 
 typedef enum _rbkit_message_fields {
   rbkit_message_field_event_type,
@@ -14,7 +20,8 @@ typedef enum _rbkit_message_fields {
   rbkit_message_field_file,
   rbkit_message_field_line,
   rbkit_message_field_size,
-  rbkit_message_field_message_counter
+  rbkit_message_field_message_counter,
+  rbkit_message_field_correlation_id
 } rbkit_message_fields;
 
 VALUE rbkit_message_fields_as_hash();